+++ web-frontend/modules/database/components/view/ViewFilter.vue
@@ -9,7 +9,7 @@
       @click="$refs.context.toggle($refs.contextLink, 'bottom', 'left', 4)"
     >
       <i class="header__filter-icon fas fa-filter"></i>
-      Filter
+      {{ filterTitle }}
     </a>
     <ViewFilterContext
       ref="context"
@@ -39,6 +39,18 @@ export default {
     view: {
       type: Object,
       required: true,
+    },
+  },
+  computed: {
+    filterTitle() {
+      const numberOfFilters = this.view.filters.length
+      if (numberOfFilters === 0) {
+        return 'Filter'
+      } else if (numberOfFilters === 1) {
+        return `${numberOfFilters} Filter`
+      } else {
+        return `${numberOfFilters} Filters`
+      }
     },
   },
   beforeMount() {