+++ web-frontend/modules/database/components/view/ViewSort.vue
@@ -9,7 +9,7 @@
       @click="$refs.context.toggle($refs.contextLink, 'bottom', 'left', 4)"
     >
       <i class="header__filter-icon fas fa-sort"></i>
-      Sort
+      {{ sortTitle }}
     </a>
     <ViewSortContext
       ref="context"
@@ -39,6 +39,18 @@ export default {
     view: {
       type: Object,
       required: true,
+    },
+  },
+  computed: {
+    sortTitle() {
+      const numberOfSortings = this.view.sortings.length
+      if (numberOfSortings === 0) {
+        return 'Sort'
+      } else if (numberOfSortings === 1) {
+        return `${numberOfSortings} Sort`
+      } else {
+        return `${numberOfSortings} Sorts`
+      }
     },
   },
 }